+++ src/Tochka.JsonRpc.ApiExplorer/ApiExplorerConstants.cs
@@ -1,0 +1,8 @@
+namespace Tochka.JsonRpc.ApiExplorer
+{
+    public static class ApiExplorerConstants
+    {
+        public const string ActionNameProperty = "actionName";
+        public const string DefaultApiVersion = "v1";
+    }
+}