+++ src/Tochka.JsonRpc.Common/JsonRpcConstants.cs
@@ -1,4 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Tochka.JsonRpc.Common
 {
@@ -21,14 +21,30 @@ namespace Tochka.JsonRpc.Common
         public static readonly object ActionResultTypeItemKey = new object();
 
         /// <summary>
+        /// Response error code. Stored in HttpContext.Items after response is written
+        /// </summary>
+        public static readonly object ResponseErrorCodeItemKey = new object();
+
+
+        /// <summary>
         /// "." Delimiter in "controller_name.action_name" to mimic "Class.Method" notation
         /// </summary>
         public static readonly string ControllerMethodSeparator = ".";
 
         /// <summary>
-        /// "id" JSON property to look for Id in single request
+        /// "id" JSON property to look for Id in single request/response
         /// </summary>
         public static readonly string IdProperty = "id";
+
+        /// <summary>
+        /// "jsonrpc" JSON property to look for version in single request/response
+        /// </summary>
+        public static readonly string JsonrpcVersionProperty = "jsonrpc";
+
+        /// <summary>
+        /// "method" JSON property to look for method in single request
+        /// </summary>
+        public static readonly string MethodProperty = "method";
 
         /// <summary>
         /// "result" JSON property to look for successful response data in single response
@@ -41,6 +57,11 @@ namespace Tochka.JsonRpc.Common
         public static readonly string ErrorProperty = "error";
 
         /// <summary>
+        /// "code" JSON property inside "error"
+        /// </summary>
+        public static readonly string ErrorCodeProperty = "code";
+
+        /// <summary>
         /// "2.0" JSON Rpc version
         /// </summary>
         public static readonly string Version = "2.0";
@@ -49,6 +70,11 @@ namespace Tochka.JsonRpc.Common
         /// "rpc." Specification restricts use of methods with this prefix
         /// </summary>
         public static readonly string ReservedMethodPrefix = "rpc.";
+
+        /// <summary>
+        /// "rpc.discover" for OpenRPC Service Discovery
+        /// </summary>
+        public static readonly string ServiceDiscoveryMethod = "rpc.discover";
 
         /// <summary>
         /// -32000 Server Error. Usually happens on unhandled Exception
@@ -70,5 +96,15 @@ namespace Tochka.JsonRpc.Common
         /// "application/json" Default content-type expected in JSON Rpc HTTP requests
         /// </summary>
         public static readonly string ContentType = "application/json";
+
+        /// <summary>
+        /// "JsonRpc" Prefix for documents like Swagger and OpenRpc
+        /// </summary>
+        public static readonly string ApiDocumentName = "jsonrpc";
+
+        /// <summary>
+        /// "REST" Swagger document name for non-JsonRpc actions
+        /// </summary>
+        public const string DefaultSwaggerDoc = "rest";
     }
 }