+++ src/Tochka.JsonRpc.Common/Models/Request/Wrappers/IRequestWrapper.cs
@@ -1,4 +1,4 @@
-﻿namespace Tochka.JsonRpc.Common.Models.Request.Wrappers
+namespace Tochka.JsonRpc.Common.Models.Request.Wrappers
 {
     public interface IRequestWrapper
     {