+++ src/Tochka.JsonRpc.Common/Models/Response/Wrappers/SingleResponseWrapper.cs
@@ -1,4 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Tochka.JsonRpc.Common.Models.Response.Wrappers
 {