+++ src/Tochka.JsonRpc.Server/Binding/IParameterBinder.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading.Tasks;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Tochka.JsonRpc.Server.Models.Binding;