+++ src/Tochka.JsonRpc.Server/Services/IResponseReader.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Tochka.JsonRpc.Common.Models.Request;