+++ src/Tochka.JsonRpc.Server/Services/JsonRpcRoutes.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.Logging;
@@ -23,6 +23,7 @@ namespace Tochka.JsonRpc.Server.Services
             log.LogTrace($"Registered route [{route}]");
         }
 
+        // TODO use TemplateMatcher
         public bool IsJsonRpcRoute(string route) => routes.Contains(route.TrimEnd('/'));
     }
 }